+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - プロパティメニューに`プロパティ名をコピー`を追加
 - `Motion@FlowType_K` に `Offset` を追加
 - `Kerning@FlowType_K` に `Effect::Parameters` を追加
+- `Blink@FlowType_K` に `Based On` を追加
 - `Island@FlowType_K` で `Time Offset::Order` によって個別オブジェクトのインデックス付与順が変わるように変更
 - キャッシュ処理の修正
 - テキストオブジェクトの情報を必要とするエフェクトをフィルタ効果やグループ制御からかけれるように修正